Update test-conditions.md
The original point of the list was to indicate which conditions are not covered by ixml.ixml.  When at the group's request I added the conditions which *are* covered, I failed to retain any distinction between conditions tested by ixml.ixml and conditions not tested by it.  I've restored the distinction by marking the latter with double bangs.  

In the meantime, of course, ixml.ixml has changed, so this list needs updating as a whole (and the process of creating it clearly needs to be automated).