A batch of 24 sample grammars with tests
The first three grammars in this batch are varying versions of the
grammar of Oberon-0, a subset of Oberon defined by Niklaus Wirth in a
1996 book on compiler construction.  The other 21 grammars are drawn
or adapted from Dick Grune and Ceriel Jacobs, Parsing Techniques.  For
each grammar an effort is made to define at least one positive and at
least one negative test; for simple grammars defining small languages
an attempt is made to provide reasonable coverage.