Minor additions and changes to error test set (#120)
* Minor additions and changes to error test set

Additional tests have been added for nonterminals which are not
NCNames and for serialization of non-XML characters.  Not all of these
tests produce errors, which makes them stick out in this set of tests,
but they are closely related to test already in this test catalog and
fit better here than elsewhere.

Several test cases with the same name in different test sets have been
renamed to reduce confusion.

* Correct accidental name duplication in errors/test-catalog.xml

* Supply missing namespace undeclarations on some outputs

Fixes the four problems found by NDW with bad default namespaces on
some asserted results.  Also copy edited a test set description and
updated the date of the test catalog.