Add another ambiguity example with A, B, C.
The ambiguity in the Battr test isn’t *really* about the marks. But
since the B nonterminals have different marks, both rules have to be
kept. I’ve modified the description of the Battr test to clarify this.

I’ve added another test that doesn’t involve marks. Here the grammar
contains two absolutely identical rules. This is ambiguous if the
processor keeps both of them and isn’t if the processor ignores
duplicated rules. Either strategy is allowed by the specification.