An attempt to reconcile the perspectives on serialization
On a careful re-reading of the iXML specification, it seems that much of the
prose uses the word “serialization” as a short hand for “constructing XML”.
There are a few places where this is made explicit and a few places where it
really does mean constructing a sequence of characters.

This PR is intended as a starting point for discussion on how we might improve
the specification’s different perspectives on serialization.